moves getPaymentMethodDetails into a util func file used in profile
